+++ auth/internal/compute/manufacturer_windows.go
